rework the ChapterFlagEnabled text
If an ordered chapter chapter is disabled it should not be part of the duration.
It is not part of the content to play. If it's enabled by a Control Track it
becomes part of the content to play and the duration can change.

That's also in the definition of ChapterFlagEnabled in ebml_matroska.xml